start thinking about Escalator
This is another game from the AisleRiot collection that I enjoyed, but which
will actually merit use of our `Distribution` for reasoning about unseen cards
(which doesn't feel as useful in Diamond Mine). Come to think of it, "Elevator"
(in which inaccessible cards are face down) is probably even more interesting
from this perspective—maybe we should do both, and try to share code!

Or maybe we should stop procrastinating with solitaire and do some useful
work??